fixed memory allocation for fake node, made some fixes and file splitting
