+++ nikonov.andrew/S1/ListNode.hpp
@@ -1,0 +1,9 @@
+namespace nikonov
+{
+  template< typename T >
+  struct ListNode
+  {
+    T data;
+    ListNode< T > * next;
+  };
+}
